Fix for the build stalling.
The tika-parsers dependency was causing maven to want to contact mvnsearch repository and this currently appears to be down. This attempt was being made for every build and the timeout is long.

This just tells maven to not use the specified repository for releases or snapshots. The ID on the repository has to match that defined in the transitive dependency.