WL-5297 Disable re-submission option on some sites
The submission sites (that are used for formal submissions) shouldn’t allow resubmissions to be enabled on an assignment as there is a formal process when a student wishes to re-submit.

We just disable the UI here and so while it’s technically possible for someone to use chrome devtools to get around this I don’t think it’s worth doing a fix for that.